requirement.txt
altgraph==0.17.4
ansicon==1.89.0
appdirs==1.4.4
blessed==1.20.0
flyingcircus==0.1.4.1
jinxed==1.3.0
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pytk==0.0.2.1
pywin32-ctypes==0.2.3
setuptools==75.6.0
setuptools-scm==8.1.0
six==1.17.0
ttkbootstrap==1.10.1
wcwidth==0.2.13
